Replaced duplicated code with a singular JsonLoader class for encoding and decoding.
