+++ Sluggo/Models/Team.swift
@@ -17,42 +17,4 @@ struct TeamRecord: Codable {
     var activated: Date?
     var deactivated: Date?
    
-    static func createFromJSON(_ string: String) -> TeamRecord? {
-        // Get data
-        guard let jsonData = string.data(using: .utf8) else {
-            print("Failed to decode provided JSON string into data for Team object intialization.")
-            return nil
-        }
-
-        // Attempt decoding
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-
-        guard let teamRec = try? decoder.decode(TeamRecord.self, from: jsonData) else {
-            print("Failed to decode JSON data into object representation for Team object initialization.")
-            return nil
-        }
-          return teamRec
-      }
-    
-    var jsonString: String? {
-        get {
-            // obtain an encoder
-            
-            let encoder = JSONEncoder()
-            encoder.dateEncodingStrategy = .iso8601
-            encoder.outputFormatting = .prettyPrinted
-            
-            // Attempt encoding
-            guard let jsonData = try? encoder.encode(self) else {
-                print("Failed to encode Team object into JSON data.")
-                return nil
-                
-            }
-            
-            // Attempt stringifying the data, this is failable, which is fine since property is optional.
-            return String(data: jsonData, encoding: .utf8)
-        }
-    }
-    
 }
